Resolve db type for datetimes (#108)

+++ server/pgsql-local/init-db.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS "User" (
 CREATE TABLE IF NOT EXISTS "Notebook" (
   nb_id UUID DEFAULT uuid_generate_v4()  PRIMARY KEY,
   name VARCHAR(256),
-  time_modified BIGINT,
+  time_modified DOUBLE PRECISION,
   language VARCHAR(256)
 );
 
@@ -25,9 +25,9 @@ CREATE TABLE IF NOT EXISTS "ActiveSession" (
   nb_id UUID REFERENCES "Notebook" (nb_id),
   uid UUID REFERENCES "User" (uid) NOT NULL,
   "connectionId" VARCHAR(64) NOT NULL,
-  time_connected BIGINT,
-  time_disconnected BIGINT,
-  last_event BIGINT,
+  time_connected DOUBLE PRECISION,
+  time_disconnected DOUBLE PRECISION,
+  last_event DOUBLE PRECISION,
   UNIQUE ("connectionId", nb_id)
 );
 
@@ -35,7 +35,7 @@ CREATE TABLE IF NOT EXISTS "Cell" (
   nb_id UUID REFERENCES "Notebook" (nb_id) NOT NULL,
   lock_held_by UUID REFERENCES "User" (uid),
   cell_id UUID DEFAULT uuid_generate_v4() PRIMARY KEY,
-  time_modified BIGINT,
+  time_modified DOUBLE PRECISION,
   contents TEXT,
   position SERIAL NOT NULL,
   cursor_pos SMALLINT,
